History Fixes

+++ MoneyWise/Models/Transaction.cs
@@ -5,7 +5,7 @@
     {
         public int  TransactionID { get; set; }
         public required int UserID { get; set; }
-        public float Amount { get; set; }
+        public decimal Amount { get; set; }
         public DateTime? created_at { get; set; }
         public DateTime? updated_at { get; set; }
     }
@@ -13,7 +13,7 @@
     // frotend display model
      public class TransactionRequest
     {
-        public float Amount { get; set; }
+        public decimal Amount { get; set; }
         public string Description { get; set; } = string.Empty;
         public string Action { get; set; } = string.Empty;
     }
